telegram_bot/requirements.txt
# CyberBot AI - Telegram Bot Requirements
# Python 3.11+ required

# Core Bot Framework
python-telegram-bot==20.7
telegram==0.0.1

# AI and NLP
openai==1.3.8
tiktoken==0.5.2

# Web Framework and API
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0

# Database
sqlite3  # Built-in with Python

# Environment and Configuration
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# News and Web Scraping
feedparser==6.0.10
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml==4.9.3

# Payment Processing
stripe==7.8.0

# Scheduling and Background Tasks
schedule==1.2.0
APScheduler==3.10.4

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Caching (Optional)
redis==5.0.1

# Monitoring and Logging
sentry-sdk==1.38.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Utilities
uuid==1.30
hashlib  # Built-in with Python
json  # Built-in with Python
asyncio  # Built-in with Python

# Development and Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Performance
aiohttp==3.9.1
asyncpg==0.29.0  # For PostgreSQL async support

# File Processing
Pillow==10.1.0  # For image processing if needed

# HTTP Client
httpx==0.25.2

# Configuration Management
pydantic==2.5.1
pydantic-settings==2.1.0

# Internationalization
babel==2.13.1

# Rate Limiting
slowapi==0.1.9

# Background Jobs
celery==5.3.4  # If using distributed task queue

# Metrics and Analytics
prometheus-client==0.19.0

